Majority of changes from review
